# fix: improve string join performance