--- a/packages/shared/test/utils.test.ts
+++ b/packages/shared/test/utils.test.ts
@@ -1,4 +1,4 @@
-import { format, generateCodeFrame, makeSymbol } from '../src/index'
+import { format, generateCodeFrame, makeSymbol, join } from '../src/index'
 
 test('format', () => {
   expect(format(`foo: {0}`, 'x')).toEqual('foo: x')
@@ -16,4 +16,44 @@ test('generateCodeFrame', () => {
 test('makeSymbol', () => {
   expect(makeSymbol('foo')).not.toEqual(makeSymbol('foo'))
   expect(makeSymbol('bar', true)).toEqual(makeSymbol('bar', true))
+})
+
+test('join', () => {
+  expect(join([])).toEqual('')
+  expect(join(['a', ','])).toEqual('a,')
+  expect(join(['a', 'b', 'c'])).toEqual('abc')
+  expect(join(['a', 'b', 'c'], ' ')).toEqual('a b c')
+  expect(
+    join(
+      [
+        'a',
+        'b',
+        'c',
+        'd',
+        'e',
+        'f',
+        'g',
+        'h',
+        'i',
+        'j',
+        'k',
+        'l',
+        'm',
+        'n',
+        'o',
+        'p',
+        'q',
+        'r',
+        's',
+        't',
+        'u',
+        'v',
+        'w',
+        'x',
+        'y',
+        'z'
+      ],
+      ' '
+    )
+  ).toEqual('a b c d e f g h i j k l m n o p q r s t u v w x y z')
 })