## Before
```sh
pnpm benchmark

compilation:

compile simple message x 365,874 ops/sec ±0.36% (96 runs sampled)
compile complex message x 54,215 ops/sec ±2.59% (95 runs sampled)

simple pattern on 1000 resources:

resolve time with core x 230 ops/sec ±1.43% (88 runs sampled)
resolve time on composition x 104 ops/sec ±0.62% (78 runs sampled)
resolve time on composition with compile cache x 205 ops/sec ±0.60% (88 runs sampled)

complex pattern on 1000 resources:

resolve time with core x 188 ops/sec ±0.40% (88 runs sampled)
resolve time on composition x 33.91 ops/sec ±0.49% (60 runs sampled)
resolve time on composition with compile cache x 169 ops/sec ±0.34% (88 runs sampled)
```

## After
```sh
pnpm benchmark

compilation:

compile simple message x 363,183 ops/sec ±0.65% (92 runs sampled)
compile complex message x 58,239 ops/sec ±0.23% (97 runs sampled)

simple pattern on 1000 resources:

resolve time with core x 236 ops/sec ±0.66% (87 runs sampled)
resolve time on composition x 106 ops/sec ±0.79% (79 runs sampled)
resolve time on composition with compile cache x 209 ops/sec ±0.33% (90 runs sampled)

complex pattern on 1000 resources:

resolve time with core x 193 ops/sec ±0.96% (87 runs sampled)
resolve time on composition x 33.24 ops/sec ±1.22% (59 runs sampled)
resolve time on composition with compile cache x 171 ops/sec ±0.30% (89 runs sampled)
```